Extract types to prevent circular imports
