+++ packages/components/src/Button/IconButton.story.tsx
@@ -27,7 +27,8 @@
 import React from 'react'
 import { Add } from '@styled-icons/material/Add'
 import { Story } from '@storybook/react/types-6-0'
-import { IconButton, IconButtonProps } from './IconButton'
+import { IconButton } from './IconButton'
+import { IconButtonProps } from './iconButtonTypes'
 
 export default {
   component: IconButton,
