+++ packages/components/src/Button/IconButton.tsx
@@ -27,111 +27,16 @@
 import some from 'lodash/some'
 import isFunction from 'lodash/isFunction'
 import styled, { css } from 'styled-components'
-import {
-  CompatibleHTMLProps,
-  reset,
-  SpaceProps,
-  space,
-  shouldForwardProp,
-  SizeLarge,
-  SizeMedium,
-  SizeSmall,
-  SizeXSmall,
-  SizeXXSmall,
-} from '@looker/design-tokens'
-import { Property } from 'csstype'
+import { reset, space, shouldForwardProp } from '@looker/design-tokens'
 import React, { forwardRef, Ref } from 'react'
-import { Placement } from '@popperjs/core'
-import { Icon, IconProps } from '../Icon'
+import { Icon } from '../Icon'
 import { useTooltip } from '../Tooltip'
 import { useForkedRef, useWrapEvent } from '../utils'
 import { VisuallyHidden } from '../VisuallyHidden'
 import { GenericButtonBase, buttonCSS } from './ButtonBase'
-import { ButtonBaseProps } from './types'
 import { iconButtonColor } from './iconButtonColor'
 import { iconButtonIconSizeMap, buttonSizeMap } from './size'
-
-interface IconButtonVariantProps {
-  /**
-   * Defines the variant or mapping of colors to style properties, like border of the button.
-   * @default false
-   */
-  outline?: boolean
-}
-
-export type IconButtonSizes =
-  | SizeXXSmall
-  | SizeXSmall
-  | SizeSmall
-  | SizeMedium
-  | SizeLarge
-
-export type ToggleColor = 'key' | 'calculation' | 'dimension' | 'measure'
-export interface IconButtonProps
-  extends Omit<CompatibleHTMLProps<HTMLButtonElement>, 'children' | 'type'>,
-    Omit<ButtonBaseProps, 'color'>,
-    IconButtonVariantProps,
-    Pick<IconProps, 'icon'>,
-    SpaceProps {
-  type?: 'button' | 'submit' | 'reset'
-  /**
-   * Refer to the keyboard expected focus behavior
-   * @private
-   */
-  focusVisible?: boolean
-  /**
-   * Display border
-   * @default false
-   */
-  outline?: boolean
-  /**
-   *  A hidden text label for the IconButton that is accessible to assistive technology
-   */
-  label: string
-  /**
-   *  Sets the size of the button
-   * @default 'xsmall'
-   */
-  size?: IconButtonSizes
-  /**
-   *  Optional round icon button variant
-   * @default 'square'
-   */
-  shape?: 'round' | 'square'
-  /**
-   * If the IconButton is in the optional toggled on or toggled off state
-   */
-  toggle?: boolean
-  /**
-   * to improve toggle's behavior this prop will update the background-color to `${toggleColor}Accent` when toggle is true
-   * @default false
-   */
-  toggleBackground?: boolean
-  /**
-   * Change icon and background color when toggled
-   * Supports 'calculation', 'dimension', 'measure' and 'key'
-   * @default 'key'
-   */
-  toggleColor?: ToggleColor
-  /**
-   * By default IconButton shows a Tooltip with the Button's label text. Setting disableTooltip will disable that behavior.
-   * @default false
-   */
-
-  tooltipDisabled?: boolean
-  /**
-   * Placement of the built-in Tooltip.
-   */
-  tooltipPlacement?: Placement
-  /**
-   * Width of the built-in Tooltip.
-   */
-  tooltipWidth?: string
-  /**
-   * Text alignment of the built-in Tooltip.
-   */
-  tooltipTextAlign?: Property.TextAlign
-}
+import { IconButtonProps } from './iconButtonTypes'
 
 export const IconButtonStyle = styled.button
   .withConfig({
