+++ packages/components/src/Button/index.ts
@@ -33,6 +33,7 @@ export * from './ButtonToggle'
 export * from './ButtonTransparent'
 export * from './IconButton'
 export * from './iconButtonColor'
+export * from './iconButtonTypes'
 export * from './MultiFunctionButton'
 export * from './types'
